pyproject.toml
[tool.poetry]
name = "ksg-nett"
version = "2023.1.1"
description = ""
authors = ["Tormod Haugland <tormod.haugland@gmail.com>", "Alexander Orvik <alexao1111@gmail.com>", "Christian De Fréne <defrenechr@me.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.13.1,!=3.14.0"
django = "^3.2.15"
pillow = "^9.0.0"
channels = "^2.4.0"
redis = "^3.5.3"
channels-redis = "^3.1.0"
django-cors-headers = "^3.2"
factory-boy = "^3.1.0"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
packaging = "^20.4"
drf-yasg2 = "^1.18.5"
weasyprint = "^54"
graphene-django = "^2.13.0"
graphene-django-cud = "^0.9.0"
django-filter = "^2.4.0"
sentry-sdk = "1.11.1"
boto3 = "^1.16.63"
django-storages = "^1.11.1"
pyjwt = "^2.3.0"
pydash = "^5.1.0"
openpyxl = "^3.0.10"
icalendar = "^5.0.1"
bleach = "^5.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.3"
pytest-django = "^3.7"
pytest-cov = "^2.8"
pre-commit = "^2.10.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | _build
  | build
)/
'''

[build-system]
requires = ["poetry>=1.2.2"]
build-backend = "poetry.masonry.api"
